Fix: Move back some thunks to slice to prevent issues
