+++ src/containers/FavouriteList.js
@@ -4,8 +4,7 @@
 import React, { useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import Favourite from '../components/Favourite';
-import { selectAllFavourites } from '../reducers/gameSlice';
-import { fetchFavourites } from '../services/favouriteCalls';
+import { selectAllFavourites, fetchFavourites } from '../reducers/gameSlice';
 
 const FavouriteList = () => {
   const dispatch = useDispatch();
