+++ src/services/favouriteCalls.js
@@ -4,24 +4,24 @@
 import axios from 'axios';
 import { createAsyncThunk } from '@reduxjs/toolkit';
 
-export const fetchFavourites = createAsyncThunk(
-  'games/fetchFavourites',
-  async (arg, { getState }) => {
-    const state = getState();
+// export const fetchFavourites = createAsyncThunk(
+//   'games/fetchFavourites',
+//   async (arg, { getState }) => {
+//     const state = getState();
 
-    try {
-      const response = await axios.get('http://127.0.0.1:3000/favourites', {
-        params: {
-          name: state.game.userName.toLowerCase(),
-        },
-      });
-      return response.data;
-    } catch (error) {
-      console.log(error);
-      throw error;
-    }
-  },
-);
+//     try {
+//       const response = await axios.get('http://127.0.0.1:3000/favourites', {
+//         params: {
+//           name: state.game.userName.toLowerCase(),
+//         },
+//       });
+//       return response.data;
+//     } catch (error) {
+//       console.log(error);
+//       throw error;
+//     }
+//   },
+// );
 
 export const setFavourite = createAsyncThunk(
   'games/setFavourite',
@@ -30,7 +30,6 @@ export const setFavourite = createAsyncThunk(
 
     try {
       const formData = new FormData();
-      formData.append('type', 'favourite');
       formData.append('name', `${state.game.userName.toLowerCase()}`);
       formData.append('game_id', `${state.game.gameInfo[0].id}`);
 
@@ -51,14 +50,18 @@ export const deleteFavourite = createAsyncThunk(
     const state = getState();
 
     try {
-      const formData = new FormData();
-      formData.append('type', 'unfavourite');
-      formData.append('name', `${state.game.userName.toLowerCase()}`);
-      formData.append('game_id', `${state.game.gameInfo[0].id}`);
+      // const formData = new FormData();
+      // formData.append('name', `${state.game.userName.toLowerCase()}`);
+      // formData.append('game_id', `${state.game.gameInfo[0].id}`);
 
-      const response = await axios.post(
-        'http://127.0.0.1:3000/favourites',
-        formData,
+      const response = await axios.delete(
+        `http://127.0.0.1:3000/favourites/${state.game.gameInfo[0].id}`,
+        {
+          params: {
+            name: state.game.userName.toLowerCase(),
+            game_id: state.game.gameInfo[0].id,
+          },
+        },
       );
 
       return response.data;
